add oshibki

+++ controllers/cards.js
@@ -11,29 +11,51 @@ module.exports.createCard = (req, res) => {
 
   Card.create({ name, link, owner: req.user._id })
     .then((card) => res.send(card))
-    .catch(() => res.status(500).send({ message: 'Произошла ошибка' }));
+    .catch((err) => {
+      if (err.name === 'ValidationError') {
+        res.status(400).send({ message: 'Переданы невалидные данные' });
+      }
+    });
 };
 
 module.exports.deleteCard = (req, res) => {
   Card.findById(req.params.cardId)
     .then((card) => {
-      Card.deleteOne(card)
-        .then(() => res.send(card))
-        .catch((err) => res.status(500).send({ message: err.name }));
+      if (!card) res.status(404).send({ message: 'Карточка c указанным _id не найдена.' });
+      else {
+        Card.deleteOne(card)
+          .then(() => res.send(card));
+      }
     })
-    .catch((err) => res.status(500).send({ message: err.name }));
+    .catch((err) => {
+      if (err.name === 'CastError') res.status(404).send({ message: 'Переданы некорректные данные при удалении карточки.' });
+    });
 };
 
 module.exports.putLike = (req, res) => {
   Card.findByIdAndUpdate(req.params.cardId, { $addToSet: { likes: req.user._id } }, { new: true })
     .populate(['owner', 'likes'])
-    .then((card) => { res.send(card); })
-    .catch((err) => res.status(500).send({ message: err.name }));
+    .then((card) => {
+      if (!card) res.status(404).send({ message: 'Карточка c указанным id не найдена' });
+      else res.send(card);
+    })
+    .catch((err) => {
+      if (err.name === 'CastError') {
+        res.status(400).send({ message: 'Переданы некорректные данные для постановки лайка' });
+      }
+    });
 };
 
 module.exports.deleteLike = (req, res) => {
   Card.findByIdAndUpdate(req.params.cardId, { $pull: { likes: req.user._id } }, { new: true })
     .populate(['owner', 'likes'])
-    .then((card) => { res.send(card); })
-    .catch((err) => res.status(500).send({ message: err.name }));
+    .then((card) => {
+      if (!card) res.status(404).send({ message: 'Карточка c указанным id не найдена' });
+      else res.send(card);
+    })
+    .catch((err) => {
+      if (err.name === 'CastError') {
+        res.status(400).send({ message: 'Переданы некорректные данные для постановки лайка' });
+      }
+    });
 };
